Split up tests into files
